+++ mirai-api-http/src/main/kotlin/net/mamoe/mirai/api/http/data/common/MessageDTO.kt
@@ -36,8 +36,14 @@ data class UnKnownMessagePacketDTO(val msg: String) : MessagePacketDTO()
 
 // Message
 @Serializable
+@SerialName("Source")
+data class MessageSourceDTO(val uid: Long) : MessageDTO()
+@Serializable
 @SerialName("At")
 data class AtDTO(val target: Long, val display: String) : MessageDTO()
+@Serializable
+@SerialName("AtAll")
+data class AtAllDTO(val target: Long = 0) : MessageDTO() // target为保留字段
 @Serializable
 @SerialName("Face")
 data class FaceDTO(val faceId: Int) : MessageDTO()
@@ -82,7 +88,9 @@ fun MessageChainDTO.toMessageChain() =
 
 @UseExperimental(ExperimentalUnsignedTypes::class)
 fun Message.toDTO() = when (this) {
+    is MessageSource -> MessageSourceDTO(messageUid)
     is At -> AtDTO(target, display)
+    is AtAll -> AtAllDTO(0L)
     is Face -> FaceDTO(id.value.toInt())
     is PlainText -> PlainDTO(stringValue)
     is Image -> ImageDTO(imageId)
@@ -93,11 +101,12 @@ fun Message.toDTO() = when (this) {
 @UseExperimental(ExperimentalUnsignedTypes::class, MiraiInternalAPI::class)
 fun MessageDTO.toMessage() = when (this) {
     is AtDTO -> At(target, display)
+    is AtAllDTO -> AtAll
     is FaceDTO -> Face(FaceId(faceId.toUByte()))
     is PlainDTO -> PlainText(text)
     is ImageDTO -> Image(imageId)
     is XmlDTO -> XMLMessage(xml)
-    is UnknownMessageDTO -> PlainText("assert cannot reach")
+    is MessageSourceDTO, is UnknownMessageDTO -> PlainText("assert cannot reach")
 }