+++ mirai-core/src/commonMain/kotlin/net.mamoe.mirai/utils/DeviceInfo.kt
@@ -11,16 +11,15 @@ package net.mamoe.mirai.utils
 
 import kotlinx.serialization.SerialId
 import kotlinx.serialization.Serializable
+import kotlinx.serialization.Transient
 import kotlinx.serialization.protobuf.ProtoBuf
-import net.mamoe.mirai.utils.cryptor.contentToString
 
 /**
  * 设备信息. 可通过继承 [SystemDeviceInfo] 来在默认的基础上修改
  */
-abstract class DeviceInfo internal constructor(
-    context: Context
-) {
-    val context: Context by context.unsafeWeakRef()
+abstract class DeviceInfo {
+    @Transient
+    abstract val context: Context
 
     abstract val display: ByteArray
     abstract val product: ByteArray
@@ -95,6 +94,45 @@ abstract class DeviceInfo internal constructor(
     }
 }
 
+@Serializable
+class DeviceInfoData(
+    override val display: ByteArray,
+    override val product: ByteArray,
+    override val device: ByteArray,
+    override val board: ByteArray,
+    override val brand: ByteArray,
+    override val model: ByteArray,
+    override val bootloader: ByteArray,
+    override val fingerprint: ByteArray,
+    override val bootId: ByteArray,
+    override val procVersion: ByteArray,
+    override val baseBand: ByteArray,
+    override val version: VersionData,
+    override val simInfo: ByteArray,
+    override val osType: ByteArray,
+    override val macAddress: ByteArray,
+    override val wifiBSSID: ByteArray?,
+    override val wifiSSID: ByteArray?,
+    override val imsiMd5: ByteArray,
+    override val imei: String,
+    override val apn: ByteArray
+) : DeviceInfo() {
+    @Transient
+    override lateinit var context: Context
+
+    @UseExperimental(ExperimentalUnsignedTypes::class)
+    override val ipAddress: ByteArray
+        get() = localIpAddress().split(".").map { it.toUByte().toByte() }.takeIf { it.size == 4 }?.toByteArray() ?: byteArrayOf()
+    override val androidId: ByteArray get() = display
+
+    @Serializable
+    class VersionData(
+        override val incremental: ByteArray = SystemDeviceInfo.Version.incremental,
+        override val release: ByteArray = SystemDeviceInfo.Version.release,
+        override val codename: ByteArray = SystemDeviceInfo.Version.codename,
+        override val sdk: Int = SystemDeviceInfo.Version.sdk
+    ) : Version
+}
 /**
  * Defaults "%4;7t>;28<fc.5*6".toByteArray()
  */