+++ mirai-core/src/jvmMain/kotlin/net/mamoe/mirai/utils/SystemDeviceInfo.kt
@@ -10,38 +10,68 @@
 package net.mamoe.mirai.utils
 
 import kotlinx.io.core.toByteArray
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.Transient
+import kotlinx.serialization.UnstableDefault
+import kotlinx.serialization.json.Json
 import net.mamoe.mirai.utils.io.getRandomByteArray
 import net.mamoe.mirai.utils.io.getRandomString
+import java.io.File
 
+/**
+ * 加载一个设备信息. 若文件不存在或为空则随机并创建一个设备信息保存.
+ */
+@UseExperimental(UnstableDefault::class)
+fun File.loadAsDeviceInfo(context: Context = ContextImpl()): DeviceInfo {
+    if (!this.exists() || this.length() == 0L) {
+        return SystemDeviceInfo(context).also {
+            this.writeText(Json.plain.stringify(SystemDeviceInfo.serializer(), it))
+        }
+    }
+    return Json.nonstrict.parse(DeviceInfoData.serializer(), this.readText()).also {
+        it.context = context
+    }
+}
+
+
+@Serializable
 @UseExperimental(ExperimentalUnsignedTypes::class)
-actual open class SystemDeviceInfo actual constructor(context: Context) : DeviceInfo(context) {
-    override val display: ByteArray get() = "MIRAI.200122.001".toByteArray()
-    override val product: ByteArray get() = "mirai".toByteArray()
-    override val device: ByteArray get() = "mirai".toByteArray()
-    override val board: ByteArray get() = "mirai".toByteArray()
-    override val brand: ByteArray get() = "mamoe".toByteArray()
-    override val model: ByteArray get() = "mirai".toByteArray()
-    override val bootloader: ByteArray get() = "unknown".toByteArray()
-    override val fingerprint: ByteArray get() = "mamoe/mirai/mirai:10/MIRAI.200122.001/5891938:user/release-keys".toByteArray()
+actual open class SystemDeviceInfo actual constructor() : DeviceInfo() {
+    actual constructor(context: Context) : this() {
+        this.context = context
+    }
+
+    @Transient
+    final override lateinit var context: Context
+
+    override val display: ByteArray = "MIRAI.200122.001".toByteArray()
+    override val product: ByteArray = "mirai".toByteArray()
+    override val device: ByteArray = "mirai".toByteArray()
+    override val board: ByteArray = "mirai".toByteArray()
+    override val brand: ByteArray = "mamoe".toByteArray()
+    override val model: ByteArray = "mirai".toByteArray()
+    override val bootloader: ByteArray = "unknown".toByteArray()
+    override val fingerprint: ByteArray = "mamoe/mirai/mirai:10/MIRAI.200122.001/${getRandomString(7, '0'..'9')}:user/release-keys".toByteArray()
     override val bootId: ByteArray = ExternalImage.generateUUID(md5(getRandomByteArray(16))).toByteArray()
-    override val procVersion: ByteArray get() = "Linux version 3.0.31-g6fb96c9 (android-build@xxx.xxx.xxx.xxx.com)".toByteArray()
-    override val baseBand: ByteArray get() = byteArrayOf()
-    override val version: DeviceInfo.Version get() = Version
-    override val simInfo: ByteArray get() = "T-Mobile".toByteArray()
-    override val osType: ByteArray get() = "android".toByteArray()
-    override val macAddress: ByteArray get() = "02:00:00:00:00:00".toByteArray()
-    override val wifiBSSID: ByteArray? get() = "02:00:00:00:00:00".toByteArray()
-    override val wifiSSID: ByteArray? get() = "<unknown ssid>".toByteArray()
-    override val imsiMd5: ByteArray get() = md5(getRandomByteArray(16))
-    override val imei: String get() = getRandomString(15, '0'..'9')
+    override val procVersion: ByteArray = "Linux version 3.0.31-${getRandomString(8)} (android-build@xxx.xxx.xxx.xxx.com)".toByteArray()
+    override val baseBand: ByteArray = byteArrayOf()
+    override val version: Version = Version
+    override val simInfo: ByteArray = "T-Mobile".toByteArray()
+    override val osType: ByteArray = "android".toByteArray()
+    override val macAddress: ByteArray = "02:00:00:00:00:00".toByteArray()
+    override val wifiBSSID: ByteArray? = "02:00:00:00:00:00".toByteArray()
+    override val wifiSSID: ByteArray? = "<unknown ssid>".toByteArray()
+    override val imsiMd5: ByteArray = md5(getRandomByteArray(16))
+    override val imei: String = getRandomString(15, '0'..'9')
     override val ipAddress: ByteArray get() = localIpAddress().split(".").map { it.toUByte().toByte() }.takeIf { it.size == 4 }?.toByteArray() ?: byteArrayOf()
     override val androidId: ByteArray get() = display
-    override val apn: ByteArray get() = "wifi".toByteArray()
+    override val apn: ByteArray = "wifi".toByteArray()
 
-    object Version : DeviceInfo.Version {
-        override val incremental: ByteArray get() = "5891938".toByteArray()
-        override val release: ByteArray get() = "10".toByteArray()
-        override val codename: ByteArray get() = "REL".toByteArray()
-        override val sdk: Int get() = 29
+    @Serializable
+    actual object Version : DeviceInfo.Version {
+        override val incremental: ByteArray = "5891938".toByteArray()
+        override val release: ByteArray = "10".toByteArray()
+        override val codename: ByteArray = "REL".toByteArray()
+        override val sdk: Int = 29
     }
 }